Refactored: uint32_t -> size_t

+++ arch/arm/drv/serial.c
@@ -5,14 +5,14 @@
 
 #ifdef CONFIG_SERIAL
 
-#include <stdint.h>
+#include <stddef.h>
 #include <sys/serial.h>
 
 #define UART0_BASE 0x1c090000
 
 void write_serial(unsigned char a)
 {
-	*(volatile uint32_t *)(UART0_BASE) = (uint32_t) a;
+	*(volatile size_t *)(UART0_BASE) = (size_t) a;
 }
 
 #endif
