+++ include/sys/tty.h
@@ -7,14 +7,15 @@
 #define TTY_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 struct fb_info {
-	uint8_t   type;
-	uint8_t   bpp;
-	uint32_t *addr; /* 64-bit field in the specification */
-	uint32_t  width;
-	uint32_t  height;
-	uint32_t  pitch;
+	uint8_t  type;
+	uint8_t  bpp;
+	size_t  *addr; /* 64-bit field in the specification */
+	size_t   width;
+	size_t   height;
+	size_t   pitch;
 };
 
 void init_console(struct fb_info framebuffer);
@@ -23,7 +24,7 @@ void disable_cursor();
 void update_cursor();
 uint16_t get_cursor_position(void);
 void printc(uint8_t ch);
-void draw_pixel(int x, int y, uint32_t color);
+void draw_pixel(int x, int y, size_t color);
 void clear_screen(void);
 void identity_map_framebuffer();
 struct fb_info *get_fb_info();
