+++ include/x86/boot/device.h
@@ -7,9 +7,9 @@
 #define BOOTDEV_H
 
 struct boot_device {
-	uint32_t biosdev;
-	uint32_t partition;
-	uint32_t sub_partition;
+	size_t biosdev;
+	size_t partition;
+	size_t sub_partition;
 } __attribute__((packed));
 
 #endif
